Added mp4 support
I write a ffconcat formatted file using the delays-and-paths data and
feed it into ffmpeg. The input images are scaled in a way to ensure the
width and height are even which is required by the libx264 codec. A crf
of 20 was found by trial and error and I used yuv420p as pixel format
because quicktime wouldn't play the videos without it.